added security to reply deletion -- checks fb user

+++ app/controllers/checkins_controller.rb
@@ -31,14 +31,18 @@ class CheckinsController < ApplicationController
 
   def destroy
 
+    fb_user = FacebookUser.find_by_nowtoken(params[:nowtoken])
+
     checkin = Checkin.where(:_id => params[:id]).first
     if checkin.nil?
       event = Event.where(:_id => params[:id]).first
       if event.nil?
         return render :text => "invalid id", :status => :error
       end
+      return render :text => "Not Authorized", :status => :error if fb_user.id != event.facebook_user_id
       event.destroy_reply(nil)
     else
+      return render :text => "Not Authorized", :status => :error if fb_user.id != checkin.facebook_user_id
       checkin.event.destroy_reply(checkin)
     end
